refactor: extract fetch from loader to services
