+++ src/api/index.ts
@@ -1,0 +1,2 @@
+export * as ImageService from "./images";
+export * as BreedService from "./breeds";
