+++ src/routes/breeds/index.ts
@@ -1,0 +1,2 @@
+export { BreedDetail, loader as breedLoader } from "./BreedDetail";
+export { Breeds, loader as breedsLoader } from "./Breeds";
